requirements.txt
functions-framework==3.*
google-cloud-storage==3.5.0
google-cloud-bigquery==3.38.0
google-cloud-aiplatform>=1.122.0
vertexai>=1.0.0
pandas>=1.5.0
db-dtypes>=1.0.0
requests>=2.31.0
flask==2.3.3